merge: improve user verification and reservation management

+++ src/main/java/com/grupo05/coworking_space/service/UserDetailsServiceImpl.java
@@ -1,5 +1,6 @@
 package com.grupo05.coworking_space.service;
 
+import com.grupo05.coworking_space.dto.ReservationDTO;
 import com.grupo05.coworking_space.dto.UserDTO;
 import com.grupo05.coworking_space.enums.ApiError;
 import com.grupo05.coworking_space.enums.Role;
@@ -16,13 +17,17 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import org.springframework.security.core.Authentication;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+
 
 /**
  * Implementación del servicio de gestión de usuarios para la autenticación y autorización.
@@ -120,7 +125,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
      */
     public UserDTO findUserById(int id) {
         User user = userRepository.findById(id).orElse(null);
-        log.info("Usuario encontrado: {}" + user.getId());
+        log.info("Usuario encontrado: {}", user.getId());
         return userMapper.convertToDTO(user);
     }
 
@@ -138,7 +143,22 @@ public class UserDetailsServiceImpl implements UserDetailsService {
             Optional<User> user = userRepository.findByUsername(username);
             if (user.isEmpty()) 
                 throw new RequestException(ApiError.RECORD_NOT_FOUND);
-            log.info("Usuario encontrado: {}" + user.get().getId());
+            log.info("Usuario encontrado: {}", user.get().getId());
+            return userMapper.convertToDTO(user.get());
+        } catch (RequestException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new RuntimeException("Error al buscar el usuario: " + e.getMessage());
+        }
+    }
+
+    public UserDTO findByUsernameAndEmail(String username,String email) {
+        try {
+            Optional<User> user = userRepository.findByUsernameAndEmail(username, email);
+
+            if (user.isEmpty()) 
+                throw new RequestException(ApiError.RECORD_NOT_FOUND);
+            log.info("Usuario encontrado: {}", user.get().getId());
             return userMapper.convertToDTO(user.get());
         } catch (RequestException e) {
             throw e;
@@ -161,15 +181,16 @@ public class UserDetailsServiceImpl implements UserDetailsService {
      */
     public UserDTO registrerUser(UserDTO user, Role role) {
         try {
-
-            if (userRepository.findByEmail(user.getEmail()).isPresent()) 
+            log.info(adminDefaultPassword);
+            if (userRepository.findByUsernameAndEmail(user.getUsername(), user.getEmail()).isPresent()) 
                 throw new RequestException(ApiError.DUPLICATE_EMAIL);
 
             User userEntity = userMapper.convertToEntity(user);
             userEntity.setPassword(passwordEncoder.encode(userEntity.getPassword()));
             userEntity.setRole(role);
             userEntity = userRepository.save(userEntity);
-            log.info("Usuario registrado: {}" + userEntity.getId());
+            log.info("Usuario registrado: {}", userEntity.getId());
+            
             return userMapper.convertToDTO(userEntity);
         } catch (RequestException e) {
             throw e;
@@ -212,7 +233,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
             // Actualizamos el token
             UserDTO loggedUserDTO = userMapper.convertToDTO(user);
             loggedUserDTO.setToken(token);
-            log.info("Usuario logeado: {}" + user.getId());
+            log.info("Usuario logeado: {}", user.getId());
             return loggedUserDTO;
         } catch (Exception e) {
             throw new RuntimeException("Error al registrar el usuario: " + e.getMessage());
@@ -228,7 +249,7 @@ public class UserDetailsServiceImpl implements UserDetailsService {
     public List<UserDTO> findAllUser() {
         try {
             List<User> user = userRepository.findAll();
-            log.info("Se han encontrado {} usuarios" + user.size());
+            log.info("Se han encontrado {} usuarios", user.size());
             return user.stream()
                     .map(userMapper::convertToDTO)
                     .collect(Collectors.toList());
@@ -251,11 +272,21 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
             if (user.isEmpty()) 
                 throw new RequestException(ApiError.RECORD_NOT_FOUND);
-
+            //-------
+            User userDB = user.get();
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            if(userDB.getRole().equals(Role.ROLE_ADMIN) && !authentication.getName().equals("admin")){
+                throw new RequestException(ApiError.AUTHENTICATION_FAILED);
+            }
+            
+            // ROLE_ADMIN
             userRepository.deleteById(id);
-            log.info("Usuario eliminado: {}" + user.get().getId());
+            log.info("Usuario eliminado: {}", user.get().getId());
             return userMapper.convertToDTO(user.get());
-        } catch (RequestException e) {
+        } catch (DataIntegrityViolationException e) {
+            log.error("Error de integridad referencial al eliminar usuario: {}", e.getMessage());
+            throw new RequestException(ApiError.ASSOCIATED_RESOURCES);
+    }    catch (RequestException e) {
             throw e;
         } catch (Exception e) {
             throw new RuntimeException("Error al eliminar el usuario: " + e.getMessage());
